implementação das mensagens para o vendedor e a criação dos endpoint de update

+++ travel/src/main/java/ifsp/travel/model/dto/HotelRequestDTO.java
@@ -26,7 +26,7 @@ public class HotelRequestDTO implements Serializable {
 
     private String location;
 
-    private Double rate;
+    private Integer rate;
 
     private Boolean favored;
